requirements_outreach.txt
# First City Foundry — Marketing Automation Dependencies

# Core outreach
requests>=2.28.0
beautifulsoup4>=4.11.0
yagmail>=0.15.0
jinja2>=3.1.0
lxml>=4.9.0
python-dotenv>=1.0.0
rich>=13.0.0
schedule>=1.2.0

# AI content generation
httpx>=0.24.0
openai>=1.0.0

# Social media posting
requests-oauthlib>=1.3.0

# Analytics and Reporting
google-analytics-data>=0.17.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.0.0

# Standard library modules (no pip install needed):
# json, csv, time, random, logging, argparse, os, datetime, dataclasses, typing
# pathlib, smtplib, email, re, urllib, xml.etree
